pyproject.toml
[project]
name = "toon-usage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai",
    "langsmith",
    "pydantic",
    "matplotlib",
    "streamlit",
    "tiktoken",
    "toon_format",
    "python-dotenv",
]
